requirements.txt
boto3>=1.9,<2.0
liac-arff>=2.4,<3.0
numpy==1.19.2
pandas==1.1.4
psutil>=5.4,<6.0
ruamel.yaml==0.16.12
scipy>=1.1,<1.5
openml==0.10.2
scikit-learn>=0.22.2
packaging
pyyaml==5.3.1
tqdm
paramiko
scp
